API

Documentation for SeqFold.

SeqFold._v! — Function

Find, store and return the minimum free energy of the structure between i and j. If i and j don't base pair, store and return INF. See: Figure 2B of Zuker, 1981

Args: seq: The sequence being folded (uppercase). i: The start index. j: The end index (1-based, inclusive). temp: The temperature in Kelvin. vcache: Free energy cache for if i and j base pair. Stores Structure. wcache: Free energy cache for lowest energy structure from i to j. Stores Structure. emap: Energy map for DNA/RNA (Energies struct).

Returns: Structure: The minimum free energy structure possible between i and j.

SeqFold._d_g — Function

Find the free energy given delta h, s and temp

Args: d_h: The enthalpy increment in kcal / mol d_s: The entropy increment in cal / mol temp_K: The temperature in Kelvin

Returns: The free energy increment in kcal / (mol x K)

SeqFold.DNA_HAIRPIN_LOOPS — Constant

Enthalpy and entropy increments for length depedence of hairpin loops

Were calculated from delta G Table 4 of SantaLucia, 2004:

Annu.Rev.Biophs.Biomol.Struct.33:415-40 doi: 10.1146/annurev.biophys.32.110601.141800 The Thermodynamics of DNA Structural Motifs SantaLucia and Hicks, 2004

For hairpins of length 3 and 4, the entropy values are looked up in the DNATRITETRA_LOOPS Dict

From formula 8-9 of the paper: An additional 1.6 delta entropy penalty if the hairpin is closed by AT

SeqFold._multi_branch — Function

Calculate a multi-branch energy penalty using a linear formula.

From Jaeger, Turner, and Zuker, 1989. Found to be better than logarithmic in Ward, et al. 2017.

Args: seq: The sequence being folded (uppercase). i: The left starting index. k: The mid-point in the search. j: The right ending index. temp: Folding temp in Kelvin. vcache: Cache of energies where V(i,j) bond. Stores Structure. wcache: Cache of min energy of substructures between W(i,j). Stores Structure. emap: Map to DNA/RNA energies (Energies struct). helix: Whether this multibranch is enclosed by a helix (V(i,j) forms a base pair).

Returns: Structure: A multi-branch structure.

SeqFold.DNA_TRI_TETRA_LOOPS — Constant

Experimental delta H and delta S for tri/tetra loops:

Supplemental Material: Annu.Rev.Biophs.Biomol.Struct.33:415-40

doi: 10.1146/annurev.biophys.32.110601.141800 The Thermodynamics of DNA Structural Motifs SantaLucia and Hicks, 2004

delta S was computed using delta G and delta H and is in cal / (K x mol)

(versus delta H in kcal / mol)

SeqFold._stack — Function

Get the free energy for a stack. Using the indexes i and j, check whether it's at the end of the sequence or internal. Then check whether it's a match or mismatch, and return. Two edge-cases are terminal mismatches and dangling ends. The energy of a dangling end is added to the energy of a pair where i XOR j is at the sequence's end.

Args: seq: The full folding sequence (uppercase). i: The start index on left side of the pair/stack. i1: The index to the right of i. j: The end index on right side of the pair/stack. j1: The index to the left of j. temp: Temperature in Kelvin. emap: Energy map for DNA/RNA (Energies struct).

Returns: Float64: The free energy of the NN pairing.

SeqFold._pair — Function

Return a stack representation, a key for the NN maps

Args: s: Sequence being folded i: leftmost index i1: index to right of i j: rightmost index j1: index to left of j

Returns: string representation of the pair

SeqFold.DNA_INTERNAL_LOOPS — Constant

Enthalpy and entropy increments for length dependence of internal loops

Were calculated from delta G Table 4 of SantaLucia, 2004:

Annu.Rev.Biophs.Biomol.Struct.33:415-40 doi: 10.1146/annurev.biophys.32.110601.141800 The Thermodynamics of DNA Structural Motifs SantaLucia and Hicks, 2004

Additional loop sizes are accounted for with the Jacobson-Stockmayer entry extrapolation formula in paper: delta G (loop-n) = delta G (loop-x) + 2.44 x R x 310.15 x ln(n / x)

Additional correction is applied for asymmetric loops in paper: delta G (asymmetry) = |length A - length B| x 0.3 (kcal / mol) where A and B are lengths of both sides of loop

SeqFold.DNA_INTERNAL_MM — Constant

Internal mismatch table (DNA):

Allawi & SantaLucia (1997), Biochemistry 36: 10581-10594

Allawi & SantaLucia (1998), Biochemistry 37: 9435-9444

Allawi & SantaLucia (1998), Biochemistry 37: 2170-2179 *

Allawi & SantaLucia (1998), Nucl Acids Res 26: 2694-2701 *

Peyret et al. (1999), Biochemistry 38: 3468-3477 *

SeqFold.DNA_TERMINAL_MM — Constant

Terminal mismatch table (DNA):

SantaLucia & Peyret (2001) Patent Application WO 01/94611

SeqFold.DNA_NN — Constant

The Thermodynamics of DNA Structural Motifs

SantaLucia and Hicks, 2004

SeqFold._hairpin — Function

Calculate the free energy of a hairpin.

Args: seq: The sequence we're folding (uppercase). i: The index of the start of hairpin. j: The index of the end of hairpin (1-based, inclusive). temp: Temperature in Kelvin. emap: Energy map for DNA/RNA (Energies struct).

Returns: Float64: The free energy increment from the hairpin structure.

SeqFold._w! — Function

Find and return the lowest free energy structure in Sij subsequence (1-based indices). This function calculates and stores results in w_cache to avoid recomputation. Figure 2B in Zuker and Stiegler, 1981.

Args: seq: The sequence being folded. i: The start index. j: The end index (1-based, inclusive). temp: The temperature in Kelvin. vcache: Free energy cache for if i and j base pair. Stores Structure. wcache: Free energy cache for lowest energy structure from i to j. Stores Structure. emap: Energy map for DNA/RNA (Energies struct).

Returns: Structure: The minimum free energy structure for the subsequence from i to j.

SeqFold.DNA_MULTIBRANCH — Constant

a, b, c, d in a linear multi-branch energy change function.

Inferred from: Supplemental Material: Annu.Rev.Biophs.Biomol.Struct.33:415-40 doi: 10.1146/annurev.biophys.32.110601.141800 The Thermodynamics of DNA Structural Motifs SantaLucia and Hicks, 2004

SeqFold._internal_loop — Function

Calculate the free energy of an internal loop.

The first and last bp of both left and right sequences are not themselves parts of the loop, but are the terminal bp on either side of it. They are needed for when there's a single internal looping bp (where just the mismatching free energies are used).

Note that both left and right sequences are in 5' to 3' direction.

This is adapted from the "Internal Loops" section of SantaLucia/Hicks, 2004.

Args: seq: The sequence we're folding (uppercase). i: The index of the start of structure on left side. i1: The index to the right of i. j: The index of the end of structure on right side. j1: The index to the left of j. temp: Temperature in Kelvin. emap: Dictionary mapping to energies for DNA/RNA (Energies struct).

Returns: Float64: The free energy associated with the internal loop.

SeqFold._j_s — Function

Estimate the free energy of length query_len based on one of length known_len.

The Jacobson-Stockmayer entry extrapolation formula is used for bulges, hairpins, etc that fall outside the 30nt upper limit for pre-calculated free-energies. See SantaLucia and Hicks (2004).

Args: querylen: Length of element without known free energy value knownlen: Length of element with known free energy value (dgx) dgx: The free energy of the element knownlen tempK: Temperature in Kelvin

Returns: Free energy for a structure of length query_len

SeqFold._trackback_energy — Function

Add energy to each structure, based on how it's W(i,j) differs from the one after.

Args: structs: The structures for whom energy is being calculated.

Returns: Vector{Structure}: Structures in the folded DNA with energy.

SeqFold._bulge — Function

Calculate the free energy associated with a bulge.

Args: seq: The full folding DNA/RNA sequence (uppercase). i: The start index of the bulge. i1: The index to the right of i. j: The end index of the bulge. j1: The index to the left of j. temp: Temperature in Kelvin. emap: Map to DNA/RNA energies (Energies struct).

Returns: Float64: The increment in free energy from the bulge.

SeqFold.DNA_DE — Constant

DNA dangling ends:

Bommarito et al. (2000), Nucl Acids Res 28: 1929-1934

SeqFold.DNA_BULGE_LOOPS — Constant

Enthalpy and entropy increments for length depedence of bulge loops

Were calculated from delta G Table 4 of SantaLucia, 2004:

Annu.Rev.Biophs.Biomol.Struct.33:415-40 doi: 10.1146/annurev.biophys.32.110601.141800 The Thermodynamics of DNA Structural Motifs SantaLucia and Hicks, 2004

For bulge loops of size 1, the intervening NN energy is used. Closing AT penalty is applied on both sides

SeqFold._min_struct — Function

Return the struct with the lowest free energy that isn't -inf (undef)

Args: structs: Structures being compared

Returns: struct: The min free energy structure

SeqFold.dg_cache — Function

Fold a nucleic acid sequence and return the estimated ΔG of each (i,j) pairing.

Args: seq: The nucleic acid sequence to fold

Keyword Args: temp: The temperature to fold at

Returns: Cache: A Vector{Vector} where each [i][j] pairing corresponds to the minimum free energy between i and j

SeqFold._traceback — Function

Traceback thru the V(i,j) and W(i,j) caches to find the structure.

For each step, get to the lowest energy W(i,j) within that block. Store the structure in W(i,j). Inc i and j. If the next structure is viable according to V(i,j), store as well. Repeat.

Args: i: The leftmost index to start searching in. j: The rightmost index to start searching in. vcache: Cache of energies where i and j bond. Stores Structure. wcache: Cache of energies/sub-structures between or with i and j. Stores Structure.

Returns: Vector{Structure}: A list of Structs in the final secondary structure.

SeqFold._cache — Function

Create caches for the wcache and vcache

The Structs is useful for gathering many possible energies between a series of (i,j) combinations.

Args: seq: The sequence to fold

Keyword Args: temp: The temperature to fold at

Returns: (w_cache, v_cache) tuple

SeqFold.SeqFold

SeqFold.DNA_BULGE_LOOPS

SeqFold.DNA_DE

SeqFold.DNA_HAIRPIN_LOOPS

SeqFold.DNA_INTERNAL_LOOPS

SeqFold.DNA_INTERNAL_MM

SeqFold.DNA_MULTIBRANCH

SeqFold.DNA_NN

SeqFold.DNA_TERMINAL_MM

SeqFold.DNA_TRI_TETRA_LOOPS

SeqFold.MeltingConditions

SeqFold.Structure

SeqFold._bulge

SeqFold._cache

SeqFold._d_g

SeqFold._hairpin

SeqFold._internal_loop

SeqFold._j_s

SeqFold._min_struct

SeqFold._multi_branch

SeqFold._pair

SeqFold._stack

SeqFold._traceback

SeqFold._trackback_energy

SeqFold._v!

SeqFold._w!

SeqFold.dg

SeqFold.dg_cache

SeqFold.dot_bracket

SeqFold.fold

SeqFold.gc_cache

SeqFold.tm

SeqFold.tm_cache